# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todolist Sorter",
    "description": "Self-hosted service that watches a Todoist project for new items and reorders them into a user-defined category sequence using an LLM. Optionally also transforms item content (typo fixes, emoji) based on per-project `additional_instructions`.\n\n**Auth.** Management endpoints require an `X-API-Key: <APP_API_KEY>` header. The webhook, OAuth callback, health probe, and `/setup` pages are public.\n\nFirst-time setup: open `/setup` in a browser.",
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/mcdax/todolist-sorter"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/webhook/{provider}": {
      "post": {
        "tags": [
          "webhook"
        ],
        "summary": "Provider webhook receiver",
        "description": "Receives webhook events from a task provider (currently `todoist`). Verifies the provider-specific signature (`X-Todoist-Hmac-SHA256` for Todoist), parses the event body, and either schedules a sort (via the debouncer) or drops the event with status `ignored` / `suppressed`. Always responds 200 with a JSON `{status: ...}` body so the provider will not retry.",
        "operationId": "receive_webhook__provider__post",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event accepted. `queued` = sort scheduled; `ignored` = unknown project / no project id in payload; `suppressed` = echo of an item the service just wrote to itself.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid JSON body."
          },
          "401": {
            "description": "Signature verification failed."
          },
          "404": {
            "description": "Unknown provider id in URL path."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}